+++ lib/extensions/geolocation.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
-import log from './logger';
+import { fs} from '@appium/support';
+import log from '../logger';
 import { exec } from 'teen_process';
-import { fs } from '@appium/support';
 
 const LYFT_SET_LOCATION = 'set-simulator-location';
 const DECIMAL_SEPARATOR_SCRIPT = `
@@ -109,6 +109,36 @@ async function setLocationWithAppleScript (sim, latitude, longitude, menu = 'Deb
   }
 }
 
-export {
-  setLocationWithLyft, setLocationWithIdb, setLocationWithAppleScript
+const extensions = {};
+
+/**
+ * Set custom geolocation parameters for the given Simulator using AppleScript.
+ *
+ * @param {string|number} latitude - The latitude value, which is going to be entered
+ *   into the corresponding edit field, for example '39,0006'.
+ * @param {string|number} longitude - The longitude value, which is going to be entered
+ *   into the corresponding edit field, for example '19,0068'.
+ * @returns {boolean} True if the given parameters have correct format and were successfully accepted.
+ * @throws {Error} If there was an error while setting the location
+ */
+extensions.setGeolocation = async function setGeolocation (latitude, longitude) {
+  const locationSetters = [
+    async () => await setLocationWithLyft(this.udid, latitude, longitude),
+    async () => await setLocationWithIdb(this.idb, latitude, longitude),
+    async () => await setLocationWithAppleScript(this, latitude, longitude, this._locationMenu),
+  ];
+
+  let lastError;
+  for (const setter of locationSetters) {
+    try {
+      await setter();
+      return true;
+    } catch (e) {
+      log.info(e.message);
+      lastError = e;
+    }
+  }
+  throw lastError;
 };
+
+export default extensions;
