+++ lib/extensions/permissions.js
@@ -1,0 +1,180 @@
+import log from '../logger';
+import _ from 'lodash';
+import { fs, util } from '@appium/support';
+import { exec } from 'teen_process';
+import path from 'path';
+
+const STATUS_UNSET = 'unset';
+const STATUS_NO = 'no';
+const STATUS_YES = 'yes';
+const WIX_SIM_UTILS = 'applesimutils';
+const SERVICES = Object.freeze({
+  calendar: 'kTCCServiceCalendar',
+  camera: 'kTCCServiceCamera',
+  contacts: 'kTCCServiceAddressBook',
+  homekit: 'kTCCServiceWillow',
+  microphone: 'kTCCServiceMicrophone',
+  photos: 'kTCCServicePhotos',
+  reminders: 'kTCCServiceReminders',
+  medialibrary: 'kTCCServiceMediaLibrary',
+  motion: 'kTCCServiceMotion',
+  health: 'kTCCServiceMSO',
+  siri: 'kTCCServiceSiri',
+  speech: 'kTCCServiceSpeechRecognition',
+});
+
+function toInternalServiceName (serviceName) {
+  if (_.has(SERVICES, _.toLower(serviceName))) {
+    return SERVICES[_.toLower(serviceName)];
+  }
+  throw new Error(`'${serviceName}' is unknown. Only the following service names are supported: ${JSON.stringify(_.keys(SERVICES))}`);
+}
+
+function formatStatus (status) {
+  return [STATUS_UNSET, STATUS_NO].includes(status) ? _.toUpper(status) : status;
+}
+
+/**
+ * Runs a command line sqlite3 query
+ *
+ * @param {string} db - Full path to sqlite database
+ * @param {string} query - The actual query string
+ * @param {...string} queryParams - The list of query parameters
+ * @returns {string} sqlite command stdout
+ */
+async function execSQLiteQuery (db, query, ...queryParams) {
+  query = query.replace(/\n+/g, ' ');
+  const queryTokens = query.split('?');
+  const formattedQuery = [];
+  queryParams
+    .map((param) => `${param}`)
+    .forEach((param, i) => {
+      formattedQuery.push(queryTokens[i]);
+      formattedQuery.push(param.replace(/'/g, "''"));
+    });
+  formattedQuery.push(queryTokens[queryTokens.length - 1]);
+
+  log.debug(`Executing SQL query "${formattedQuery.join('')}" on '${db}'`);
+  try {
+    return (await exec('sqlite3', ['-line', db, formattedQuery.join('')])).stdout;
+  } catch (err) {
+    throw new Error(
+      `Cannot execute SQLite query "${formattedQuery.join('')}" to '${db}'. ` +
+      `Original error: ${err.stderr}`
+    );
+  }
+}
+
+async function execWix (args) {
+  try {
+    await fs.which(WIX_SIM_UTILS);
+  } catch (e) {
+    throw new Error(
+      `${WIX_SIM_UTILS} binary has not been found in your PATH. ` +
+      `Please install it ('brew tap wix/brew && brew install wix/brew/applesimutils') to ` +
+      `be able to change application permissions`
+    );
+  }
+
+  log.debug(`Executing: ${WIX_SIM_UTILS} ${util.quote(args)}`);
+  try {
+    const {stdout} = await exec(WIX_SIM_UTILS, args);
+    log.debug(`Command output: ${stdout}`);
+    return stdout;
+  } catch (e) {
+    throw new Error(`Cannot execute "${WIX_SIM_UTILS} ${util.quote(args)}". Original error: ${e.stderr || e.message}`);
+  }
+}
+
+/**
+ * Sets permissions for the given application
+ *
+ * @param {string} bundleId - bundle identifier of the target application.
+ * @param {Object} permissionsMapping - An object, where keys ar  service names
+ * and values are corresponding state values. See https://github.com/wix/AppleSimulatorUtils
+ * for more details on available service names and statuses.
+ * @throws {Error} If there was an error while changing permissions.
+ */
+async function setAccess (bundleId, permissionsMapping) {
+  const permissionsArg = _.toPairs(permissionsMapping)
+    .map((x) => `${x[0]}=${formatStatus(x[1])}`)
+    .join(',');
+  return await execWix([
+    '--byId', this.udid,
+    '--bundle', bundleId,
+    '--setPermissions', permissionsArg,
+  ]);
+}
+
+/**
+ * Retrieves the current permission status for the given service and application.
+ *
+ * @param {string} bundleId - bundle identifier of the target application.
+ * @param {string} serviceName - the name of the service. Should be one of
+ * `SERVICES` keys.
+ * @param {string} sharedResourcesDir - the path to Simulator `data` root
+ * @returns {string} - The current status: yes/no/unset
+ * @throws {Error} If there was an error while retrieving permissions.
+ */
+async function getAccess (bundleId, serviceName, sharedResourcesDir) {
+  const internalServiceName = toInternalServiceName(serviceName);
+  const dbPath = path.resolve(sharedResourcesDir, 'Library', 'TCC', 'TCC.db');
+  for (const [sqlValue, status] of [['0', STATUS_NO], ['1', STATUS_YES]]) {
+    const count = await execSQLiteQuery(
+      dbPath,
+      `SELECT count(*) FROM 'access' WHERE client='?' AND allowed=? AND service='?'`,
+      bundleId, sqlValue, internalServiceName
+    );
+    if (parseInt(count.split('=')[1], 10) > 0) {
+      return status;
+    }
+  }
+  return STATUS_UNSET;
+}
+
+
+const extensions = {};
+
+/**
+ * Sets the particular permission to the application bundle. See
+ * https://github.com/wix/AppleSimulatorUtils for more details on
+ * the available service names and statuses.
+ *
+ * @param {string} bundleId - Application bundle identifier.
+ * @param {string} permission - Service name to be set.
+ * @param {string} value - The desired status for the service.
+ * @throws {Error} If there was an error while changing permission.
+ */
+extensions.setPermission = async function setPermission (bundleId, permission, value) {
+  await this.setPermissions(bundleId, {[permission]: value});
+};
+
+/**
+ * Sets the permissions for the particular application bundle.
+ *
+ * @param {string} bundleId - Application bundle identifier.
+ * @param {Object} permissionsMapping - A mapping where kays
+ * are service names and values are their corresponding status values.
+ * See https://github.com/wix/AppleSimulatorUtils
+ * for more details on available service names and statuses.
+ * @throws {Error} If there was an error while changing permissions.
+ */
+extensions.setPermissions = async function setPermissions (bundleId, permissionsMapping) {
+  log.debug(`Setting access for '${bundleId}': ${JSON.stringify(permissionsMapping, null, 2)}`);
+  await setAccess(bundleId, permissionsMapping);
+};
+
+/**
+ * Retrieves current permission status for the given application bundle.
+ *
+ * @param {string} bundleId - Application bundle identifier.
+ * @param {string} serviceName - One of available service names.
+ * @throws {Error} If there was an error while retrieving permissions.
+ */
+extensions.getPermission = async function getPermission (bundleId, serviceName) {
+  const result = await getAccess(bundleId, serviceName, this.getDir());
+  log.debug(`Got ${serviceName} access status for '${bundleId}': ${result}`);
+  return result;
+};
+
+export default extensions;
