+++ lib/simulator-xcode-8.js
@@ -1,22 +1,15 @@
 import _ from 'lodash';
 import log from './logger';
 import { exec } from 'teen_process';
-import {
-  setLocationWithLyft, setLocationWithIdb, setLocationWithAppleScript
-} from './geolocation';
-import {
-  MOBILE_SAFARI_BUNDLE_ID, SAFARI_STARTUP_TIMEOUT, launchApp,
-  getDeveloperRoot, activateApp, SIMULATOR_APP_NAME
-} from './utils';
+import { activateApp, SIMULATOR_APP_NAME } from './utils';
 import path from 'path';
 import { getPath as getXcodePath } from 'appium-xcode';
-import { fs, tempDir, mkdirp, plist, timing, util } from '@appium/support';
-import B from 'bluebird';
+import { fs, plist, timing } from '@appium/support';
 import AsyncLock from 'async-lock';
 import { retryInterval, waitForCondition } from 'asyncbox';
 import { EventEmitter } from 'events';
 import Simctl from 'node-simctl';
-import { generateDefaultsCommandArgs } from './defaults-utils';
+import extensions from './extensions/index';
 
 /*
  * This event is emitted as soon as iOS Simulator
@@ -126,7 +119,7 @@ class SimulatorXcode8 extends EventEmitter {
    */
   async getPlatformVersion () {
     if (!this._platformVersion) {
-      let {sdk} = await this.stat();
+      const {sdk} = await this.stat();
       this._platformVersion = sdk;
     }
     return this._platformVersion;
@@ -157,54 +150,6 @@ class SimulatorXcode8 extends EventEmitter {
    */
   getLogDir () {
     return path.resolve(process.env.HOME, 'Library', 'Logs', 'CoreSimulator', this.udid);
-  }
-
-  /**
-   * Install valid .app package on Simulator.
-   *
-   * @param {string} app - The path to the .app package.
-   */
-  async installApp (app) {
-    return await this.simctl.installApp(app);
-  }
-
-  /**
-   * Returns user installed bundle ids which has 'bundleName' in their Info.Plist as 'CFBundleName'
-   * @param {string} bundleName - The bundle name of the application to be checked.
-   * @return {array<string>} - The list of bundle ids which have 'bundleName'
-   */
-  async getUserInstalledBundleIdsByBundleName (bundleName) {
-    const appsRoot = path.resolve(this.getDir(), 'Containers', 'Bundle', 'Application');
-    // glob all Info.plist from simdir/data/Containers/Bundle/Application
-    const infoPlists = await fs.glob('*/*.app/Info.plist', {
-      cwd: appsRoot,
-      nosort: true,
-      strict: false,
-      absolute: true,
-    });
-    if (_.isEmpty(infoPlists)) {
-      return [];
-    }
-
-    const bundleInfoPromises = [];
-    for (const infoPlist of infoPlists) {
-      bundleInfoPromises.push((async () => {
-        try {
-          return await plist.parsePlistFile(infoPlist);
-        } catch (ign) {}
-      })());
-    }
-    const bundleInfos = await B.all(bundleInfoPromises);
-    const bundleIds = bundleInfos.filter(({ CFBundleName }) => CFBundleName === bundleName);
-    if (_.isEmpty(bundleIds)) {
-      return [];
-    }
-
-    log.debug(
-      `The simulator has ${util.pluralize('bundle', bundleIds.length, true)} which ` +
-      `have '${bundleName}' as their 'CFBundleName': ${JSON.stringify(bundleIds)}`
-    );
-    return bundleIds;
   }
 
   /**
@@ -242,7 +187,7 @@ class SimulatorXcode8 extends EventEmitter {
     }
 
     const { lastBootedAt } = await plist.parsePlistFile(devicePlist);
-    return !!lastBootedAt;
+    return !lastBootedAt;
   }
 
   /**
@@ -359,27 +304,6 @@ class SimulatorXcode8 extends EventEmitter {
   }
 
   /**
-   * Scrub (delete the preferences and changed files) the particular application on Simulator.
-   *
-   * @param {string} bundleId - Bundle identifier of the application.
-   */
-  async scrubCustomApp (bundleId) {
-    const appDataRoot = await this.simctl.getAppContainer(bundleId, 'data');
-    const appFiles = await fs.glob('**/*', {
-      cwd: appDataRoot,
-      nosort: true,
-      strict: false,
-      nodir: true,
-      absolute: true,
-    });
-    log.info(`Found ${appFiles.length} ${bundleId} app ${util.pluralize('file', appFiles.length, false)} to scrub`);
-    if (_.isEmpty(appFiles)) {
-      return;
-    }
-    await B.all(appFiles.map((p) => fs.rimraf(p)));
-  }
-
-  /**
    * @typedef {Object} ShutdownOptions
    * @property {?number|string} timeout The number of milliseconds to wait until
    * Simulator is shut down completely. No wait happens if the timeout value is not set
@@ -416,195 +340,6 @@ class SimulatorXcode8 extends EventEmitter {
    */
   async delete () {
     await this.simctl.deleteDevice();
-  }
-
-  /**
-   * Updates variious Safari settings. Simulator must be booted in order to for it
-   * to success.
-   *
-   * @param {object} updates An object containing Safari settings to be updated.
-   * The list of available setting names and their values could be retrived by
-   * changing the corresponding Safari settings in the UI and then inspecting
-   * 'Library/Preferences/com.apple.mobilesafari.plist' file inside of
-   * com.apple.mobilesafari app container.
-   * The full path to the Mobile Safari's container could be retrieved from
-   * `xcrun simctl get_app_container <sim_udid> com.apple.mobilesafari data`
-   * command output.
-   * Use the `xcrun simctl spawn <sim_udid> defaults read <path_to_plist>` command
-   * to print the plist content to the Terminal.
-   */
-  async updateSafariSettings (updates) {
-    if (_.isEmpty(updates)) {
-      return false;
-    }
-
-    const containerRoot = await this.simctl.getAppContainer(MOBILE_SAFARI_BUNDLE_ID, 'data');
-    const plistPath = path.join(containerRoot, 'Library', 'Preferences', 'com.apple.mobilesafari.plist');
-    return await this.updateSettings(plistPath, updates);
-  }
-
-  /**
-   * Updates Reduce Motion setting state.
-   *
-   * @param {boolean} reduceMotion Whether to enable or disable the setting.
-   */
-  async setReduceMotion (reduceMotion) {
-    return await this.updateSettings('com.apple.Accessibility', {
-      ReduceMotionEnabled: Number(reduceMotion)
-    });
-  }
-
-  /**
-   * Updates Reduce Transparency setting state.
-   *
-   * @param {boolean} reduceTransparency Whether to enable or disable the setting.
-   */
-  async setReduceTransparency (reduceTransparency) {
-    return await this.updateSettings('com.apple.Accessibility', {
-      EnhancedBackgroundContrastEnabled: Number(reduceTransparency)
-    });
-  }
-
-  /**
-   * Allows to update Simulator preferences in runtime.
-   *
-   * @param {string} domain The name of preferences domain to be updated,
-   * for example, 'com.apple.Preferences' or 'com.apple.Accessibility' or
-   * full path to a plist file on the local file system.
-   * @param {object} updates Mapping of keys/values to be updated
-   * @returns {boolean} True if settings were actually changed
-   */
-  async updateSettings (domain, updates) {
-    if (_.isEmpty(updates)) {
-      return false;
-    }
-
-    const argChunks = generateDefaultsCommandArgs(updates);
-    await B.all(argChunks.map((args) => this.simctl.spawnProcess([
-      'defaults', 'write', domain, ...args
-    ])));
-    return true;
-  }
-
-  /**
-   * Sets UI appearance style.
-   * This function can only be called on a booted simulator.
-   *
-   * @since Xcode SDK 11.4
-   */
-  async setAppearance (/* value */) { // eslint-disable-line require-await
-    throw new Error(`Xcode SDK '${this.xcodeVersion}' is too old to set UI appearance`);
-  }
-
-  /**
-   * Gets the current UI appearance style
-   * This function can only be called on a booted simulator.
-   *
-   * @since Xcode SDK 11.4
-   */
-  async getAppearance () { // eslint-disable-line require-await
-    throw new Error(`Xcode SDK '${this.xcodeVersion}' is too old to get UI appearance`);
-  }
-
-  // eslint-disable-next-line require-await
-  async configureLocalization () {
-    throw new Error(`Xcode SDK '${this.xcodeVersion}' is too old to configure the Simulator locale`);
-  }
-
-  /**
-   * Clean up the directories for mobile Safari.
-   *
-   * @param {boolean} keepPrefs - Whether to keep Safari preferences from being deleted.
-   */
-  async cleanSafari (keepPrefs = true) {
-    try {
-      if (await this.isRunning()) {
-        await this.simctl.terminateApp(MOBILE_SAFARI_BUNDLE_ID);
-      }
-    } catch (ign) {}
-
-    log.debug('Cleaning mobile safari data files');
-    if (await this.isFresh()) {
-      log.info(
-        'Could not find Safari support directories to clean out old data. ' +
-        'Probably there is nothing to clean out'
-      );
-      return;
-    }
-
-    let safariData = null;
-    try {
-      safariData = await this.simctl.getAppContainer(MOBILE_SAFARI_BUNDLE_ID, 'data');
-    } catch (ign) {};
-    if (!safariData) {
-      log.info(
-        'Could not find Safari support directories to clean out old ' +
-        'data. Probably there is nothing to clean out'
-      );
-      return;
-    }
-    const libraryDir = path.resolve(safariData, 'Library');
-    const filesToDelete = [
-      ['Caches', '*'],
-      ['Image Cache', '*'],
-      ['WebKit', MOBILE_SAFARI_BUNDLE_ID, '*'],
-      ['WebKit', 'GeolocationSites.plist'],
-      ['WebKit', 'LocalStorage', '*.*'],
-      ['Safari', '*'],
-      ['Cookies', '*.binarycookies'],
-      ['..', 'tmp', MOBILE_SAFARI_BUNDLE_ID, '*'],
-    ];
-    const deletePromises = filesToDelete.map((p) => fs.rimraf(path.resolve(libraryDir, ...p)));
-    if (!keepPrefs) {
-      deletePromises.push(fs.rimraf(path.resolve(libraryDir, 'Preferences', '*.plist')));
-    }
-    await B.all(deletePromises);
-  }
-
-  /**
-   * Uninstall the given application from the current Simulator.
-   *
-   * @param {string} bundleId - The buindle ID of the application to be removed.
-   */
-  async removeApp (bundleId) {
-    await this.simctl.removeApp(bundleId);
-  }
-
-  /**
-   * Perform Simulator caches cleanup.
-   *
-   * @param {...string} folderNames - The names of Caches subfolders to be cleaned.
-   *   Non-accessible/non-existing subfolders will be skipped.
-   *   All existing subfolders under Caches will be deleted if this parameter is omitted.
-   * @returns {number} The count of cleaned cache items.
-   *   Zero is returned if no items were matched for cleanup (either not accessible or not directories).
-   */
-  async clearCaches (...folderNames) {
-    const cachesRoot = path.resolve(this.getDir(), 'Library', 'Caches');
-    if (!(await fs.hasAccess(cachesRoot))) {
-      log.debug(`Caches root at '${cachesRoot}' does not exist or is not accessible. Nothing to do there`);
-      return 0;
-    }
-
-    let itemsToRemove = folderNames.length ? folderNames : (await fs.readdir(cachesRoot));
-    itemsToRemove = itemsToRemove.map((x) => path.resolve(cachesRoot, x));
-    if (folderNames.length) {
-      itemsToRemove = await B.filter(itemsToRemove, (x) => fs.hasAccess(x));
-    }
-    itemsToRemove = await B.filter(itemsToRemove, async (x) => (await fs.stat(x)).isDirectory());
-    if (!itemsToRemove.length) {
-      log.debug(`No Simulator cache items for cleanup were matched in '${cachesRoot}'`);
-      return 0;
-    }
-
-    log.debug(`Matched ${util.pluralize('simulator cache item', itemsToRemove.length, true)} ` +
-      `for cleanup: ${itemsToRemove}`);
-    try {
-      await B.all(itemsToRemove, (x) => fs.rimraf(x));
-    } catch (e) {
-      log.warn(`Got an exception while cleaning Simulator caches: ${e.message}`);
-    }
-    return itemsToRemove.length;
   }
 
   /**
@@ -660,109 +395,6 @@ class SimulatorXcode8 extends EventEmitter {
     });
   }
 
-  //region Keychains Interaction
-  /**
-   * Create the backup of keychains folder.
-   * The previously created backup will be automatically
-   * deleted if this method was called twice in a row without
-   * `restoreKeychains` being invoked.
-   *
-   * @returns {boolean} True if the backup operation was successfull.
-   */
-  async backupKeychains () {
-    if (!await fs.exists(this.keychainPath)) {
-      return false;
-    }
-
-    const backupPath = await tempDir.path({
-      prefix: `keychains_backup_${Math.floor((1 + Math.random()) * 0x10000).toString(16).substring(1)}`,
-      suffix: '.zip',
-    });
-    const zipArgs = [
-      '-r', backupPath,
-      `${this.keychainPath}${path.sep}`
-    ];
-    log.debug(`Creating keychains backup with 'zip ${zipArgs.join(' ')}' command`);
-    await exec('zip', zipArgs);
-    if (_.isString(this._keychainsBackupPath) && await fs.exists(this._keychainsBackupPath)) {
-      await fs.unlink(this._keychainsBackupPath);
-    }
-    this._keychainsBackupPath = backupPath;
-    return true;
-  }
-
-  /**
-   * Restore the previsouly created keychains backup.
-   *
-   * @param {?string|Array<string>} excludePatterns - The list
-   * of file name patterns to be excluded from restore. The format
-   * of each item should be the same as '-x' option format for
-   * 'unzip' utility. This can also be a comma-separated string,
-   * which is going be transformed into a list automatically,
-   * for example: '*.db*,blabla.sqlite'
-   * @returns {boolean} If the restore opration was successful.
-   * @throws {Error} If there is no keychains backup available for restore.
-   */
-  async restoreKeychains (excludePatterns = []) {
-    if (!_.isString(this._keychainsBackupPath) || !await fs.exists(this._keychainsBackupPath)) {
-      throw new Error(`The keychains backup archive does not exist. ` +
-                      `Are you sure it was created before?`);
-    }
-
-    if (_.isString(excludePatterns)) {
-      excludePatterns = excludePatterns.split(',').map((x) => x.trim());
-    }
-    const isServerRunning = await this.isRunning();
-    let plistPath;
-    if (isServerRunning) {
-      plistPath = path.resolve(await this.getLaunchDaemonsRoot(), 'com.apple.securityd.plist');
-      if (!await fs.exists(plistPath)) {
-        throw new Error(`Cannot clear keychains because '${plistPath}' does not exist`);
-      }
-      await this.simctl.spawnProcess(['launchctl', 'unload', plistPath]);
-    }
-    try {
-      await fs.rimraf(this.keychainPath);
-      await mkdirp(this.keychainPath);
-      const unzipArgs = [
-        '-o', this._keychainsBackupPath,
-        ...(_.flatMap(excludePatterns.map((x) => ['-x', x]))),
-        '-d', '/'
-      ];
-      log.debug(`Restoring keychains with 'unzip ${unzipArgs.join(' ')}' command`);
-      await exec('unzip', unzipArgs);
-      await fs.unlink(this._keychainsBackupPath);
-      this._keychainsBackupPath = null;
-    } finally {
-      if (isServerRunning && plistPath) {
-        await this.simctl.spawnProcess(['launchctl', 'load', plistPath]);
-      }
-    }
-    return true;
-  }
-
-  /**
-   * Clears Keychains for the particular simulator in runtime (there is no need to stop it).
-   *
-   * @throws {Error} If keychain cleanup has failed.
-   */
-  async clearKeychains () {
-    const plistPath = path.resolve(await this.getLaunchDaemonsRoot(), 'com.apple.securityd.plist');
-    if (!await fs.exists(plistPath)) {
-      throw new Error(`Cannot clear keychains because '${plistPath}' does not exist`);
-    }
-    await this.simctl.spawnProcess(['launchctl', 'unload', plistPath]);
-    try {
-      if (await fs.exists(this.keychainPath)) {
-        await fs.rimraf(this.keychainPath);
-        await mkdirp(this.keychainPath);
-      }
-    } finally {
-      await this.simctl.spawnProcess(['launchctl', 'load', plistPath]);
-    }
-  }
-  //endregion
-
   /**
    * @typedef {Object} ProcessInfo
    * @property {number} pid The actual process identifier.
@@ -815,76 +447,6 @@ class SimulatorXcode8 extends EventEmitter {
       });
     }
     return result;
-  }
-
-  /**
-   * Sets the particular permission to the application bundle. See
-   * https://github.com/wix/AppleSimulatorUtils for more details on
-   * the available service names and statuses.
-   *
-   * @param {string} bundleId - Application bundle identifier.
-   * @param {string} permission - Service name to be set.
-   * @param {string} value - The desired status for the service.
-   * @throws {Error} If there was an error while changing permission.
-   */
-  async setPermission (bundleId, permission, value) {
-    await this.setPermissions(bundleId, {[permission]: value});
-  }
-
-  /**
-   * Sets the permissions for the particular application bundle.
-   *
-   * @param {string} bundleId - Application bundle identifier.
-   * @param {Object} permissionsMapping - A mapping where kays
-   * are service names and values are their corresponding status values.
-   * See https://github.com/wix/AppleSimulatorUtils
-   * for more details on available service names and statuses.
-   * @throws {Error} If there was an error while changing permissions.
-   */
-  async setPermissions (bundleId, permissionsMapping) {
-    log.debug(`Setting access for '${bundleId}': ${JSON.stringify(permissionsMapping, null, 2)}`);
-    await this.permissions.setAccess(bundleId, permissionsMapping);
-  }
-
-  /**
-   * Retrieves current permission status for the given application bundle.
-   *
-   * @param {string} bundleId - Application bundle identifier.
-   * @param {string} serviceName - One of available service names.
-   * @throws {Error} If there was an error while retrieving permissions.
-   */
-  async getPermission (bundleId, serviceName) {
-    const result = await this.permissions.getAccess(bundleId, serviceName);
-    log.debug(`Got ${serviceName} access status for '${bundleId}': ${result}`);
-    return result;
-  }
-
-  /**
-   * Adds the given certificate into the Trusted Root Store on the simulator.
-   * The simulator must be shut down in order for this method to work properly.
-   *
-   * @param {string} payload the content of the PEM certificate
-   * @returns {boolean} `true` if the certificate has been successfully installed
-   * or `false` if it has already been there
-   */
-  // eslint-disable-next-line require-await
-  async addCertificate () {
-    throw new Error(`Xcode SDK '${this.xcodeVersion}' is too old add certificates`);
-  }
-
-  /**
-   * Simulates push notification delivery
-   *
-   * @since Xcode SDK 11.4
-   */
-  async pushNotification (/* payload */) { // eslint-disable-line require-await
-    throw new Error(`Xcode SDK '${this.xcodeVersion}' is too old to push notifications`);
-  }
-
-  async getLaunchDaemonsRoot () {
-    const devRoot = await getDeveloperRoot();
-    return path.resolve(devRoot,
-      'Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator.sdk/System/Library/LaunchDaemons');
   }
 
   /**
@@ -950,30 +512,6 @@ class SimulatorXcode8 extends EventEmitter {
   }
 
   /**
-   * Verify whether the particular application is installed on Simulator.
-   * @override
-   *
-   * @param {string} bundleId - The bundle id of the application to be checked.
-   * @return {boolean} True if the given application is installed.
-   */
-  async isAppInstalled (bundleId) {
-    try {
-      const appContainer = await this.simctl.getAppContainer(bundleId);
-      return appContainer.endsWith('.app');
-    } catch (err) {
-      // get_app_container subcommand fails for system applications,
-      // so we try the hidden appinfo subcommand, which prints correct info for
-      // system/hidden apps
-      try {
-        const info = await this.simctl.appInfo(bundleId);
-        return info.includes('ApplicationType');
-      } catch (e) {
-        return false;
-      }
-    }
-  }
-
-  /**
    * @return {number} The max number of milliseconds to wait until Simulator booting is completed.
    */
   get startupTimeout () {
@@ -991,125 +529,10 @@ class SimulatorXcode8 extends EventEmitter {
     await this.simctl.startBootMonitor({timeout: startupTimeout});
     this.emit(BOOT_COMPLETED_EVENT);
   }
+}
 
-  /**
-   * Open the given URL in mobile Safari browser.
-   * The browser will be started automatically if it is not running.
-   *
-   * @param {string} url - The URL to be opened.
-   */
-  async openUrl (url) {
-    if (!await this.isRunning()) {
-      throw new Error(`Tried to open '${url}', but Simulator is not in Booted state`);
-    }
-    const timer = new timing.Timer().start();
-    try {
-      await launchApp(this.simctl, MOBILE_SAFARI_BUNDLE_ID, SAFARI_STARTUP_TIMEOUT);
-      await this.simctl.openUrl(url);
-    } catch (err) {
-      throw new Error(`Safari could not open '${url}' after ${timer.getDuration().asSeconds.toFixed(3)}s. ` +
-        `Original error: ${err.stderr || err.message}`);
-    }
-    log.debug(`Safari successfully opened '${url}' in ${timer.getDuration().asSeconds.toFixed(3)}s`);
-  }
-
-  /**
-   * Perform Shake gesture on Simulator window.
-   */
-  async shake () {
-    log.info(`Performing shake gesture on ${this.udid} Simulator`);
-    await this.simctl.spawnProcess([
-      'notifyutil',
-      '-p', 'com.apple.UIKit.SimulatorShake'
-    ]);
-  }
-
-  /**
-   * Get the current state of Biometric Enrollment feature.
-   *
-   * @returns {boolean} Either true or false
-   * @throws {Error} If Enrollment state cannot be determined
-   */
-  async isBiometricEnrolled () {
-    const output = await this.executeUIClientScript(`
-      tell application "System Events"
-        tell process "Simulator"
-          set dstMenuItem to menu item "Toggle Enrolled State" of menu 1 of menu item "Touch ID" of menu 1 of menu bar item "Hardware" of menu bar 1
-          set isChecked to (value of attribute "AXMenuItemMarkChar" of dstMenuItem) is "✓"
-        end tell
-      end tell
-    `);
-    log.debug(`Touch ID enrolled state: ${output}`);
-    return _.isString(output) && output.trim() === 'true';
-  }
-
-  /**
-   * Enrolls biometric (TouchId, FaceId) feature testing in Simulator UI client.
-   *
-   * @param {boolean} isEnabled - Defines whether biometric state is enabled/disabled
-   * @throws {Error} If the enrolled state cannot be changed
-   */
-  async enrollBiometric (isEnabled = true) {
-    await this.executeUIClientScript(`
-      tell application "System Events"
-        tell process "Simulator"
-          set dstMenuItem to menu item "Toggle Enrolled State" of menu 1 of menu item "Touch ID" of menu 1 of menu bar item "Hardware" of menu bar 1
-          set isChecked to (value of attribute "AXMenuItemMarkChar" of dstMenuItem) is "✓"
-          if ${isEnabled ? 'not ' : ''}isChecked then
-            click dstMenuItem
-          end if
-        end tell
-      end tell
-    `);
-  }
-
-  /**
-   * Sends a notification to match/not match the touch id.
-   *
-   * @param {?boolean} shouldMatch [true] - Set it to true or false in order to emulate
-   * matching/not matching the corresponding biometric
-   */
-  async sendBiometricMatch (shouldMatch = true) {
-    await this.executeUIClientScript(`
-      tell application "System Events"
-        tell process "Simulator"
-          set dstMenuItem to menu item "${shouldMatch ? 'Matching Touch' : 'Non-matching Touch'}" of menu 1 of menu item "Touch ID" of menu 1 of menu bar item "Hardware" of menu bar 1
-          click dstMenuItem
-        end tell
-      end tell
-    `);
-  }
-
-  /**
-   * Set custom geolocation parameters for the given Simulator using AppleScript.
-   *
-   * @param {string|number} latitude - The latitude value, which is going to be entered
-   *   into the corresponding edit field, for example '39,0006'.
-   * @param {string|number} longitude - The longitude value, which is going to be entered
-   *   into the corresponding edit field, for example '19,0068'.
-   * @returns {boolean} True if the given parameters have correct format and were successfully accepted.
-   * @throws {Error} If there was an error while setting the location
-   */
-  async setGeolocation (latitude, longitude) {
-    const locationSetters = [
-      async () => await setLocationWithLyft(this.udid, latitude, longitude),
-      async () => await setLocationWithIdb(this.idb, latitude, longitude),
-      async () => await setLocationWithAppleScript(this, latitude, longitude, this._locationMenu),
-    ];
-
-    let lastError;
-    for (const setter of locationSetters) {
-      try {
-        await setter();
-        return true;
-      } catch (e) {
-        log.info(e.message);
-        lastError = e;
-      }
-    }
-    throw lastError;
-  }
-
+for (const [cmd, fn] of _.toPairs(extensions)) {
+  SimulatorXcode8.prototype[cmd] = fn;
 }
 
 export default SimulatorXcode8;
